Adds games list page
Implements a page to display a list of available games
fetched from the API.

The page displays a loading indicator while fetching data,
an error message if the load fails, and a list of games,
if any, that can be clicked to navigate to the game page.

Registers a named HttpClient "Api" targeting the backend
API service.