Removes auto-formatting during build
Removes the auto-formatting functionality during the build process.
This was previously enabled via a Directory.Build.targets file
that automatically formatted code when changes were detected.
This is being removed likely because formatting is now handled
differently or is no longer desired as part of the build process.

Relates to JAF-44